+++ src/nirfsg/examples/nirfsg_script.py
@@ -28,11 +28,11 @@ def example(resource_name, options, frequency, power_level, number_of_samples):
 
 def _main(argsv):
     parser = argparse.ArgumentParser(description='Generates a signal based on the script provided.', formatter_class=argparse.ArgumentDefaultsHelpFormatter)
-    parser.add_argument('-n', '--resource-name', default='5841', help='Resource name of the NI RF signal generator.')
+    parser.add_argument('-n', '--resource-name', default='', required=True, help='Resource name of the NI RF signal generator.')
     parser.add_argument('-f', '--frequency', default=1e9, type=float, help='Frequency in Hz.')
     parser.add_argument('-p', '--power-level', default=-10.0, type=float, help='Power level in dBm.')
     parser.add_argument('-s', '--number-of-samples', default=1000, type=int, help='Number of samples.')
-    parser.add_argument('-op', '--option-string', default='Simulate=1, DriverSetup=Model:5841', type=str, help='Option string for the session.')
+    parser.add_argument('-op', '--option-string', default='', type=str, help='Option string for the session.')
     args = parser.parse_args(argsv)
     example(args.resource_name, args.option_string, args.frequency, args.power_level, args.number_of_samples)
 
@@ -42,12 +42,12 @@ def main():
 
 
 def test_example():
-    options = "Simulate=1, DriverSetup=Model:5841"
-    example('5841', options, 1e9, -10.0, 1000)
+    options = {'simulate': True, 'driver_setup': {'Model': '5841', }, }
+    example('simulated5841', options, 1e9, -10.0, 1000)
 
 
 def test_main():
-    cmd_line = ['--resource-name', '5841', '--frequency', '1e9', '--power-level', '-10', '--number-of-samples', '1000', '--option-string', 'Simulate=1, DriverSetup=Model:5841']
+    cmd_line = ['--resource-name', 'simulated5841', '--frequency', '1e9', '--power-level', '-10', '--number-of-samples', '1000', '--option-string', 'Simulate=1, DriverSetup=Model:5841']
     _main(cmd_line)